refactor: unify appointment and block forms
