+++ src/app/components/dashboard-component/dashboard-component.html
@@ -556,10 +556,11 @@
   *ngIf="showAppointmentModal"
   [startDate]="selectedDate!"
   [appointment]="selectedAppointment"
+  [mode]="formMode"
   (onSave)="handleSaveAppointment($event)"
+  (onSaveBlock)="handleSaveTimeBlock($event)"
   (onCancel)="closeAllModals()"
->
-</app-appointment-form>
+></app-appointment-form>
 
 <app-appointment-details-modal
   *ngIf="showAppointmentDetailsModal"
@@ -568,11 +569,3 @@
   (onCancel)="closeAllModals()"
 ></app-appointment-details-modal>
 
-<app-time-block-form
-  *ngIf="showTimeBlockModal"
-  [startDate]="selectedDate!"
-  [timeBlock]="null"
-  [onSave]="handleSaveTimeBlock.bind(this)"
-  (onCancel)="closeAllModals()"
->
-</app-time-block-form>
